Made small edits to the CSS.
Added Transitions to the menu buttons in the index.html and options.html files.

Swapped out Baloo-Da font for Century Gothic.

Changed the "Wii Menu" and "Start" Button to more accurately reflect the actual Wii Menu.

Temporarily commented out the Wii Menu music so it does not play on every refresh. Will be added back in a later time.