Big Changes!
Been a bit since the last update, but here's what's been done:

- Started utilizing React.js(alongside Electron) for the project. This will allow for the project to be more organized and easier to manage.
- Due to the above, the project has been restructured to fit the new React.js format, with large portions of the code being rewritten, or are in the process of being rewritten.
- Added an Info Box to the main menu, which will display information about the selected game when hovering over a tile.
- Added Single Page Navigation to most pages, allowing for music to persist between pages.
- Changed the Healthy and Safety Warning to it's own page instead of a video
- Options Menu is currently being reworked to more closely match the original Wii Settings menu.

- KNOWN ISSUES -
- The Options Menu is currently not functional.
- When refreshing most pages, with the exception of Warning.js, the music stops.
- The new pages are yet to be properly optimized for various screen resolutions.
- The Tiles in the main menu will not play the animation for clicking on a tile, when clicking on a tile.
- Baloo Da font is currently not being used, but will be put back in a later date.
- Unable to add box art and other media to tiles.

Moving to React.js is certainly a big change for me, especially since it's my first time using it. It's a bit of a learning curve, but it's been a fun experience so far.
I already know Vue.js, so it's not too hard to pick up, but it's still a bit of a challenge. I'm excited to see where this project goes from here!

Updates from here on may be a bit slower, as I am incredibly close to graduating from College, and it's quite the busy time for me.

Have a great day!

-mightyowl866