Tile's now have images again! Created a file that handles the tile configuration, binding a game's "ID" to an image. I'll be expanding on this more in the future, but for now, it's a good start. Also added a couple of images to the public/assets folder.
Since the move to React, there is still a lot of work ahead.
The Options menu is still a WIP, and the Game Management menu is still in the works.

Currently, a new bug has appeared where the tile's images are not being loaded in properly.
But this should hopefully be an easy fix.